feat(admin): add article management functionalityImplement new article management features including controllers, services, and necessary DTOs and mappers. Refactor existing TagMapper and ArticleMapper for improved data handling. The commit also removes MyBatis logging configuration to clean up the logs.feat(admin): add article update and detail query functionality
- Implement updateArticle and findArticleDetail methods in ArticleController.- Add corresponding service and mapper implementations for article update and detail query.
- Refactor ArticleMapper and ArticleDetailMapper for improved data manipulation.
- Include new DTOs UpdateArticleReqDto and FindArticleDetailReqDto.

feat(admin): optimize article API endpoints and update date type- Refactor delete and page GET methods with specific paths for better clarity.
- Replace Java.util.Date with java.time.LocalDateTime in ArticleVo.
- Ensure all lines in the commit message are under 100 characters.

fix(admin): prevent deletion of categories and tags in useAdd checks in CategoryServiceImpl and TagServiceImpl to prevent the deletion of
categories and

+++ weblog-moudle-admin/src/main/java/com/lsj/weblog/admin/service/impl/CateGoryServiceImpl.java
@@ -2,6 +2,7 @@ package com.lsj.weblog.admin.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.lsj.weblog.admin.mapper.ArticleMapper;
 import com.lsj.weblog.admin.mapper.CategoryMapper;
 import com.lsj.weblog.admin.model.dto.AddCategoryReqDto;
 import com.lsj.weblog.admin.model.dto.FindCategoryPageReqDto;
@@ -28,6 +29,8 @@ public class CateGoryServiceImpl implements CateGoryService {
 
 
     private final CategoryMapper categoryMapper;
+
+    private final ArticleMapper articleMapper;
 
 
     @Override
@@ -65,6 +68,12 @@ public class CateGoryServiceImpl implements CateGoryService {
     public void deleteCategory(IdRequestDto idRequestDto) {
         if (idRequestDto.getId() == null) {
             throw new BizExecption(ResponseCodeEnum.VALIDATION_ERROR);
+        }
+        // 判断是否被文章使用
+        int articleNums = articleMapper.selectCountByCategoryId(idRequestDto.getId());
+
+        if (articleNums != 0) {
+            throw new BizExecption(ResponseCodeEnum.CATEGORY_USED_ERROR);
         }
 
         int nums = categoryMapper.deleteById(idRequestDto.getId());
